Moved other pom related class to the newly created package
